+++ lib/services/android-project-service.ts
@@ -10,7 +10,9 @@ import fs = require("fs");
 import os = require("os");
 
 class AndroidProjectService implements IPlatformProjectService {
-	private SUPPORTED_TARGETS = ["android-17", "android-18", "android-19", "android-21"]; // forbidden for now: "android-MNC"
+	private static MIN_SUPPORTED_VERSION = 17;
+	private SUPPORTED_TARGETS = ["android-17", "android-18", "android-19", "android-21", "android-22"]; // forbidden for now: "android-MNC"
+	private static ANDROID_TARGET_PREFIX = "android";
 	private static RES_DIRNAME = "res";
 	private static VALUES_DIRNAME = "values";
 	private static VALUES_VERSION_DIRNAME_PREFIX = AndroidProjectService.VALUES_DIRNAME + "-v";
@@ -48,7 +50,8 @@ class AndroidProjectService implements IPlatformProjectService {
 				],
 				frameworkFilesExtensions: [".jar", ".dat", ".so"],
 				configurationFileName: "AndroidManifest.xml",
-				configurationFilePath: path.join(this.$projectData.platformsDir, "android", "AndroidManifest.xml")
+				configurationFilePath: path.join(this.$projectData.platformsDir, "android", "AndroidManifest.xml"),
+				mergeXmlConfig: [{ "nodename": "manifest", "attrname": "*" }]
 			};
 		}
 
@@ -67,7 +70,9 @@ class AndroidProjectService implements IPlatformProjectService {
 	public createProject(projectRoot: string, frameworkDir: string): IFuture<void> {
 		return (() => {
 			this.$fs.ensureDirectoryExists(projectRoot).wait();
-			let newTarget = this.getLatestValidAndroidTarget(frameworkDir).wait();
+			
+			let newTarget = this.getAndroidTarget(frameworkDir).wait();
+			this.$logger.trace(`Using Android SDK '${newTarget}'.`);
 			let versionNumber = _.last(newTarget.split("-"));
 			if(this.$options.symlink) {
 				this.copyResValues(projectRoot, frameworkDir, versionNumber).wait();
@@ -101,8 +106,22 @@ class AndroidProjectService implements IPlatformProjectService {
 			this.$fs.createDirectory(resDestinationDir).wait();
 			let versionDirName = AndroidProjectService.VALUES_VERSION_DIRNAME_PREFIX + versionNumber;
 			let directoriesToCopy = [AndroidProjectService.VALUES_DIRNAME];
-			if(this.$fs.exists(path.join(resSourceDir, versionDirName)).wait()) {
-				directoriesToCopy.push(versionDirName);
+			let directoriesInResFolder = this.$fs.readDirectory(resSourceDir).wait();
+			let integerFrameworkVersion = parseInt(versionNumber);
+			let versionDir = _.find(directoriesInResFolder, dir => dir === versionDirName) || 
+				_(directoriesInResFolder)
+				.map(dir => { 
+					return {
+						dirName: dir,
+						sdkNum: parseInt(dir.substr(AndroidProjectService.VALUES_VERSION_DIRNAME_PREFIX.length))
+					}
+				})
+				.filter(dir => dir.dirName.match(AndroidProjectService.VALUES_VERSION_DIRNAME_PREFIX) && dir.sdkNum && (!integerFrameworkVersion || (integerFrameworkVersion >= dir.sdkNum)))
+				.max(dir => dir.sdkNum)
+				.dirName;
+
+			if(versionDir) {
+				directoriesToCopy.push(versionDir);
 			}
 
 			this.copy(resDestinationDir, resSourceDir, directoriesToCopy.join(" "), "-R").wait();
@@ -331,17 +350,37 @@ class AndroidProjectService implements IPlatformProjectService {
 		}
 	}
 
+	private getAndroidTarget(frameworkDir: string): IFuture<string> {
+		return ((): string => { 
+			let newTarget = this.$options.sdk ? `${AndroidProjectService.ANDROID_TARGET_PREFIX}-${this.$options.sdk}` : this.getLatestValidAndroidTarget(frameworkDir).wait();
+			if(!_.contains(this.SUPPORTED_TARGETS, newTarget)) {
+				let versionNumber = parseInt(_.last(newTarget.split("-")));
+				if(versionNumber && (versionNumber < AndroidProjectService.MIN_SUPPORTED_VERSION)) {
+					this.$errors.failWithoutHelp(`The selected target SDK ${newTarget} is not supported. You should target at least ${AndroidProjectService.MIN_SUPPORTED_VERSION}.`);
+				}
+
+				if(!_.contains(this.getInstalledTargets().wait(), newTarget)) {
+					this.$errors.failWithoutHelp(`You have selected to use ${newTarget}, but it is not currently installed.`+
+						' Run \"android\" from your command-line to install/update any missing SDKs or tools.');
+				}
+				this.$logger.warn(`The selected Android target '${newTarget}' is not verified as supported. Some functionality may not work as expected.`);
+			}
+
+			return newTarget;
+		}).future<string>()();
+	}
+
 	private getLatestValidAndroidTarget(frameworkDir: string): IFuture<string> {
 		return (() => {
 			let validTarget = this.getTarget(frameworkDir).wait();
 			let installedTargets = this.getInstalledTargets().wait();
 
 			// adjust to the latest available version
-			var	newTarget = _(this.SUPPORTED_TARGETS).sort().findLast(supportedTarget => _.contains(installedTargets, supportedTarget));
+			var newTarget = _(this.SUPPORTED_TARGETS).sort().findLast(supportedTarget => _.contains(installedTargets, supportedTarget));
 			if (!newTarget) {
-				this.$errors.fail("Please install Android target %s. Make sure you have the latest Android tools installed as well." +
-					" Run \"android\" from your command-line to install/update any missing SDKs or tools.",
-					validTarget.split('-')[1]);
+				this.$errors.failWithoutHelp(`Could not find supported Android target. Please install one of the following: ${this.SUPPORTED_TARGETS.join(", ")}.` + 
+					" Make sure you have the latest Android tools installed as well." + 
+					' Run "android" from your command-line to install/update any missing SDKs or tools.')
 			}
 
 			return newTarget;